# swagger/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/grupo-estudio/{_id}": {
            "get": {
                "operationId": "GrupoEstudioController_getById",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "grupo-estudio"
                ]
            },
            "put": {
                "operationId": "GrupoEstudioController_put",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GrupoEstudioDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "grupo-estudio"
                ]
            },
            "delete": {
                "operationId": "GrupoEstudioController_delete",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "grupo-estudio"
                ]
            }
        },
...
        "/horario-semestre-grupo-estudio": {
            "post": {
                "operationId": "HorarioSemestreGrupoEstudioController_post",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HorarioSemestreGrupoEstudioDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "horario-semestre-grupo-estudio"
                ]
            },
            "get": {
                "operationId": "HorarioSemestreGrupoEstudioController_getAll",
                "parameters": [
                    {
                        "name": "query",
                        "required": false,
                        "in": "query",
                        "description": "query - Filter. e.g. col1:v1,col2:v2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fields",
                        "required": false,
                        "in": "query",
                        "description": "fields - Fields returned. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortby",
                        "required": false,
                        "in": "query",
                        "description": "sortby - Sorted-by fields. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "required": false,
                        "in": "query",
                        "description": "order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "description": "limit - Limit the size of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offset",
                        "required": false,
                        "in": "query",
                        "description": "offset - Start position of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "populate",
                        "required": false,
                        "in": "query",
                        "description": "populate - show subqueries. Must be a boolean",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "horario-semestre-grupo-estudio"
                ]
            }
        },
        "/horario-semestre-grupo-estudio/{_id}": {
            "get": {
                "operationId": "HorarioSemestreGrupoEstudioController_getById",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "horario-semestre-grupo-estudio"
                ]
            },
            "put": {
                "operationId": "HorarioSemestreGrupoEstudioController_put",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HorarioSemestreGrupoEstudioDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "horario-semestre-grupo-estudio"
                ]
            },
            "delete": {
                "operationId": "HorarioSemestreGrupoEstudioController_delete",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "horario-semestre-grupo-estudio"
                ]
            }
        },
...
        "/horario-semestre/{_id}": {
            "get": {
                "operationId": "HorarioSemestreController_getById",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "horario-semestre"
                ]
            },
            "put": {
                "operationId": "HorarioSemestreController_put",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HorarioSemestreDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "horario-semestre"
                ]
            },
            "delete": {
                "operationId": "HorarioSemestreController_delete",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "horario-semestre"
                ]
            }
        },
...
        "/horario/{_id}": {
            "get": {
                "operationId": "HorarioController_getById",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "horario"
                ]
            },
            "put": {
                "operationId": "HorarioController_put",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HorarioDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "horario"
                ]
            },
            "delete": {
                "operationId": "HorarioController_delete",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "horario"
                ]
            }
        }
    },
    "info": {
        "title": "Horarios_crud",
        "description": "API CRUD para la gestion de horarios",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
...
            "EstadoCreacionSemestreDto": {
                "type": "object",
                "properties": {
                    "Nombre": {
                        "type": "string"
                    },
                    "Descripcion": {
                        "type": "string"
                    },
                    "CodigoAbreviacion": {
                        "type": "string"
                    },
                    "Activo": {
                        "type": "boolean"
                    },
                    "FechaCreacion": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "FechaModificacion": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "Nombre",
                    "Descripcion",
                    "CodigoAbreviacion",
                    "Activo",
                    "FechaCreacion",
                    "FechaModificacion"
                ]
            },
            "EstadoCreacionDto": {
                "type": "object",
                "properties": {
                    "Nombre": {
                        "type": "string"
                    },
                    "Descripcion": {
                        "type": "string"
                    },
                    "CodigoAbreviacion": {
                        "type": "string"
                    },
                    "Activo": {
                        "type": "boolean"
                    },
                    "FechaCreacion": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "FechaModificacion": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "Nombre",
                    "Descripcion",
                    "CodigoAbreviacion",
                    "Activo",
                    "FechaCreacion",
                    "FechaModificacion"
                ]
            },
            "GrupoEstudioDto": {
                "type": "object",
                "properties": {
                    "CodigoProyecto": {
                        "type": "string"
                    },
                    "IndicadorGrupo": {
                        "type": "string"
                    },
                    "CuposGrupos": {
                        "type": "number"
                    },
                    "EspaciosAcademicos": {
                        "type": "object"
                    },
                    "ProyectoAcademicoId": {
                        "type": "string"
                    },
                    "PlanEstudiosId": {
                        "type": "string"
                    },
                    "SemestreId": {
                        "type": "string"
                    },
                    "Activo": {
                        "type": "boolean"
                    },
                    "FechaCreacion": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "FechaModificacion": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "CodigoProyecto",
                    "IndicadorGrupo",
                    "CuposGrupos",
                    "EspaciosAcademicos",
                    "ProyectoAcademicoId",
                    "PlanEstudiosId",
                    "SemestreId",
                    "Activo",
                    "FechaCreacion",
                    "FechaModificacion"
                ]
            },
            "HorarioDto": {
                "type": "object",
                "properties": {
                    "Nombre": {
                        "type": "string"
                    },
                    "CodigoAbreviacion": {
                        "type": "string"
                    },
                    "Codigo": {
                        "type": "string"
                    },
                    "ProyectoAcademicoId": {
                        "type": "number"
                    },
                    "PlanEstudioId": {
                        "type": "number"
                    },
                    "Semestres": {
                        "type": "number"
                    },
                    "PeriodoId": {
                        "type": "number"
                    },
                    "EstadoCreacionId": {
                        "$ref": "#/components/schemas/EstadoCreacionDto"
                    },
                    "Observacion": {
                        "type": "string"
                    },
                    "Activo": {
                        "type": "boolean"
                    },
                    "FechaCreacion": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "FechaModificacion": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "Nombre",
                    "CodigoAbreviacion",
                    "Codigo",
                    "ProyectoAcademicoId",
                    "PlanEstudioId",
                    "Semestres",
                    "PeriodoId",
                    "EstadoCreacionId",
                    "Observacion",
                    "Activo",
                    "FechaCreacion",
                    "FechaModificacion"
                ]
            },
            "HorarioSemestreDto": {
                "type": "object",
                "properties": {
                    "Nombre": {
                        "type": "string"
                    },
                    "CodigoAbreviacion": {
                        "type": "string"
                    },
                    "SemestreId": {
                        "type": "number"
                    },
                    "PeriodoId": {
                        "type": "number"
                    },
                    "HorarioId": {
                        "$ref": "#/components/schemas/HorarioDto"
                    },
                    "EstadoCreacionSemestreId": {
                        "$ref": "#/components/schemas/EstadoCreacionSemestreDto"
                    },
                    "Observacion": {
                        "type": "string"
                    },
                    "Activo": {
                        "type": "boolean"
                    },
                    "FechaCreacion": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "FechaModificacion": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "Nombre",
                    "CodigoAbreviacion",
                    "SemestreId",
                    "PeriodoId",
                    "HorarioId",
                    "EstadoCreacionSemestreId",
                    "Observacion",
                    "Activo",
                    "FechaCreacion",
                    "FechaModificacion"
                ]
            },
...
            "HorarioSemestreGrupoEstudioDto": {
                "type": "object",
                "properties": {
                    "HorarioSemestreId": {
                        "$ref": "#/components/schemas/HorarioSemestreDto"
                    },
                    "GrupoEstudioId": {
                        "$ref": "#/components/schemas/GrupoEstudioDto"
                    },
                    "Activo": {
                        "type": "boolean"
                    },
                    "FechaCreacion": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "FechaModificacion": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "HorarioSemestreId",
                    "GrupoEstudioId",
                    "Activo",
                    "FechaCreacion",
                    "FechaModificacion"
                ]
            }
        }
    }
}